lazy be like

+++ src/Sensor.cpp
@@ -7,7 +7,7 @@ bool Sensor::ready() { return this->status > 0; }
 
 Sensor::SensorMesureData Sensor::__read(int index, size_t count,
                                         SensorMesureData *datas) {
-  if (index > count) {
+  if (size_t(index) > count) {
     return 0;
   }
   return datas[index];
@@ -16,7 +16,7 @@ Sensor::SensorMesureData Sensor::__read(int index, size_t count,
 Sensor::SensorMesureData Sensor::__average(
     int last, int index, size_t count,
     CircularBuffer<SensorMesureData, 20>buffers[]) {
-  if (index > count) {
+  if (size_t(index) > count) {
     return 0;
   }
   SensorMesureData data = 0;
@@ -25,7 +25,7 @@ Sensor::SensorMesureData Sensor::__average(
     last = buffers[index].size();
   }
 
-  for (size_t i = 0; i < last; i++) {
+  for (size_t i = 0; i < size_t(last); i++) {
     data = data + buffers[index][i];
   }
   return data / float(last);
